+++ example/__mocks__/async-storage.js
@@ -1,0 +1,1 @@
+export default from '@react-native-async-storage/async-storage/jest/async-storage-mock';
