+++ example/src/screens/InitialScreen.tsx
@@ -1,8 +1,6 @@
 import { BackgroundAnimation } from '@components/BackgroundAnimation';
 import { Typo } from '@components/Typo';
 import { CardButton } from '@components/buttons/CardButton';
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import * as Sentry from '@sentry/react-native';
 import * as Application from 'expo-application';
 import React from 'react';
 import { View, StyleSheet, Image, Pressable, Alert } from 'react-native';
@@ -10,20 +8,14 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import { useVideoroomState } from 'src/VideoroomContext';
 
 export const InitialScreen = ({ navigation }) => {
-  const { setIsDevMode, isDevMode } = useVideoroomState();
+  const { isDevMode, setSavedIsDevMode } = useVideoroomState();
 
   const toggleDevMode = async () => {
     const updatedIsDevMode = !isDevMode;
     Alert.alert(
       'Dev mode is now '.concat(updatedIsDevMode ? 'enabled' : 'disabled')
     );
-    setIsDevMode(updatedIsDevMode);
-    try {
-      const jsonValue = JSON.stringify(updatedIsDevMode);
-      await AsyncStorage.setItem('isDevMode', jsonValue);
-    } catch (err) {
-      Sentry.captureException(err);
-    }
+    setSavedIsDevMode(updatedIsDevMode);
   };
 
   return (
